FileAuthor

+++ Library.Admin/Controllers/ResourceController.cs
@@ -22,9 +22,12 @@ namespace Library.Admin.Controllers
         private readonly IEducationalProgramService _educationalProgramService;
         private readonly ISpecialtyService _specialtyService;
         private readonly IAuthorService _authorService;
-        public ResourceController(IFileService fileService, ILanguageService languageService, ICategoryService categoryService,
-                                    IFileTypeService fileTypeService,IEducationalProgramService educationalProgramService,
-                                    ISpecialtyService specialtyService,IAuthorService authorService)
+        private readonly IFileAuthorService _fileAuthorService;
+
+        public ResourceController(IFileService fileService, ILanguageService languageService,
+            ICategoryService categoryService,
+            IFileTypeService fileTypeService, IEducationalProgramService educationalProgramService,
+            ISpecialtyService specialtyService, IAuthorService authorService, IFileAuthorService fileAuthorService)
         {
             _fileService = fileService;
             _languageService = languageService;
@@ -33,6 +36,7 @@ namespace Library.Admin.Controllers
             _educationalProgramService = educationalProgramService;
             _specialtyService = specialtyService;
             _authorService = authorService;
+            _fileAuthorService = fileAuthorService;
         }
 
         #region Resources
@@ -130,6 +134,7 @@ namespace Library.Admin.Controllers
                     authorList.Add(new SelectListItem() { Text = $"{author.LastName} {author.FirstName} {author.FatherName}", Value = $"{author.Id}" });
                 }
 
+
                 viewModel.CategoryList = new SelectList(categories.Data, "Id", "Name", "Seç");
                 viewModel.LanguageList = new SelectList(languages.Data, "Id", "Name", "Seç");
                 viewModel.FileTypeList = new SelectList(fileTypes.Data, "Id", "Name", "Seç");
@@ -141,7 +146,10 @@ namespace Library.Admin.Controllers
                 var file = await _fileService.Get(id);
 
                 viewModel.File = file.Data;
-                
+
+                var authorIds = await _fileAuthorService.GetAllFileAuthors(viewModel.File.Id);
+                viewModel.Members = authorIds.Data;
+
                 return PartialView(viewModel);
             }
             return new NotFoundResult();
@@ -156,6 +164,13 @@ namespace Library.Admin.Controllers
             {
                 try
                 {
+                    FileAuthorDto fileAuthorDto = new FileAuthorDto()
+                    {
+                        AuthorIds = new List<int>(viewModel.Members),
+                        FileId = viewModel.File.Id
+
+                    };
+
                     if (viewModel.File.Id == 0)
                     {
                         viewModel.File.FilePath = UniqueNameGenerator.UniqueFilePathGenerator(viewModel.AddedFile.FileName);
@@ -192,7 +207,8 @@ namespace Library.Admin.Controllers
 
                             var result = await _fileService.Add(accessToken, fileDto);
 
-                            if (!result.Success)
+                            var resultAuthor = await _fileAuthorService.AddAllFileAuthors(accessToken, fileAuthorDto);
+                            if (!result.Success && !resultAuthor.Success)
                             {
                                 await DeleteFileFromFileSystem(Path.Combine(DefaultPath.OriginalDefaultFilePath, viewModel.File.FilePath));
                                 await DeleteFileFromFileSystem(Path.Combine(DefaultPath.OriginalDefaultPhotoPath, viewModel.File.PhotoPath));
@@ -249,8 +265,12 @@ namespace Library.Admin.Controllers
                                 };
 
                                 var result = await _fileService.Update(accessToken, fileDto);
-
-                                if (!result.Success)
+                                if (result.Success)
+                                {
+                                    await _fileAuthorService.DeleteFileAuthor(accessToken, viewModel.File.Id);
+                                    await _fileAuthorService.AddAllFileAuthors(accessToken, fileAuthorDto);
+                                }
+                                else
                                 {
                                     await DeleteFileFromFileSystem(Path.Combine(DefaultPath.OriginalDefaultFilePath, oldFilePath));
                                     viewModel.File.FilePath = oldFilePath;
@@ -298,7 +318,12 @@ namespace Library.Admin.Controllers
 
                                 var result = await _fileService.Update(accessToken, fileDto);
 
-                                if (!result.Success)
+                                if (result.Success)
+                                {
+                                    await _fileAuthorService.DeleteFileAuthor(accessToken, viewModel.File.Id);
+                                    await _fileAuthorService.AddAllFileAuthors(accessToken, fileAuthorDto);
+                                }
+                                else
                                 {
                                     await DeleteFileFromFileSystem(Path.Combine(DefaultPath.OriginalDefaultFilePath, oldFilePath));
                                     viewModel.File.FilePath = oldFilePath;
@@ -344,7 +369,12 @@ namespace Library.Admin.Controllers
 
                                 var result = await _fileService.Update(accessToken, fileDto);
 
-                                if (!result.Success)
+                                if (result.Success)
+                                {
+                                    await _fileAuthorService.DeleteFileAuthor(accessToken, viewModel.File.Id);
+                                    await _fileAuthorService.AddAllFileAuthors(accessToken, fileAuthorDto);
+                                }
+                                else
                                 {
                                     await DeleteFileFromFileSystem(Path.Combine(DefaultPath.OriginalDefaultPhotoPath, oldPhotoPath));
                                     viewModel.File.PhotoPath = oldPhotoPath;
@@ -379,7 +409,9 @@ namespace Library.Admin.Controllers
                                 LastModified = viewModel.File.LastModified
                             };
 
-                            var result2 = await _fileService.Update(accessToken, fileDto);
+                            await _fileService.Update(accessToken, fileDto);
+                            await _fileAuthorService.DeleteFileAuthor(accessToken, viewModel.File.Id);
+                            await _fileAuthorService.AddAllFileAuthors(accessToken, fileAuthorDto);
                         }
                     }
                     TempData["Message"] = "Operation successfully";
